InstrumentDaemon class
added multiple inheritance
